move helper functions to accessor_utils
